+++ src/renderer/Renderer.js
@@ -153,7 +153,7 @@ export default class Renderer {
         }
 
         // Assume that all dataframes of a renderLayer share the same metadata
-        const metadata = dataframes.length ? dataframes[0].metadata : {};
+        const metadata = dataframes.length ? dataframes[0].metadata : null;
 
         viewportExpressions.forEach(expr => expr._resetViewportAgg(metadata));
 
@@ -226,15 +226,11 @@ export default class Renderer {
         return dataframe.getFeature(index);
     }
 
-    renderLayer (renderLayer) {
+    renderLayer (renderLayer, drawMetadata) {
         const dataframes = renderLayer.getActiveDataframes();
         const viz = renderLayer.viz;
         const gl = this.gl;
         const aspect = this._getAspect();
-        const drawMetadata = {
-            zoom: gl.drawingBufferHeight / (this._zoom * 1024 * (window.devicePixelRatio || 1)),
-            scale: gl.drawingBufferHeight / (this._zoom * 512 * (window.devicePixelRatio || 1))
-        };
 
         this._runViewportAggregations(renderLayer);