+++ CHANGELOG.md
@@ -8,9 +8,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Add new `clusterCount()` expression
 - Support feature-dependant arrays in combination with ramp
+- Add support to `.blendTo` with String API expressions
+- Support symbol and symbolPlacement in features
+
+### Changed
+ - zoom() returns the current zoom level in the typical logarithmic form, returning the same value as Mapbox GL Map.getZoom() method
+- Unary and Binary operators are case insensitive.
 
 ### Fixed
 - Add default export to allow `import carto from '@carto/carto-vl';`
+- Fix `blendTo` with `circle` SVG
 
 ## [0.7.0] - 2018-08-24
 ### Added